Fixed more warnings.
git-svn-id: http://winbolo.googlecode.com/svn/trunk@68 f5f8f8c0-f940-0410-bb6b-693536db5c03

+++ winbolo/src/bolo/bolo_map.c
@@ -1309,7 +1309,7 @@ long mapPrepareRun(map *value, bmapRun *run, BYTE *xPos, BYTE *yPos) {
 
   *xPos = run->endx = x;
 	*yPos = run->y    = y;
-	return(run->datalen = nibble_data - (BYTE*)run);
+	return (run->datalen = (BYTE) (nibble_data - (BYTE*)run));
 }
 
 /*********************************************************
